feat: Add stub room client
We'll use this for intra-service communication later on.

The code probably won't compile in this commit because the tests are failing because NewService expects "room client" to be passed as another argument. The next commit will fix that. The reason I split them up is so that this commit is easier to understand.